# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Citadel service","description":"HTTP service for working with Security Operations Center resources","version":"1.0"},"servers":[{"url":"http://localhost:8000/citadel","description":"citadel contains the vulnerability, risk, incident, and intel services."}],"paths":{"/incidents":{"get":{"tags":["incidents"],"summary":"list all incidents","operationId":"incidents#list all","parameters":[{"name":"limit","in":"query","description":"Limit the number of results","allowEmptyValue":true,"schema":{"type":"integer","description":"Limit the number of results","example":864,"minimum":0,"maximum":1000},"example":750}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Incident"},"example":[{"Permissions":"analyst","affected_customers":["Quis ipsam.","Voluptatem eum.","Vero aliquam quisquam veritatis quia quis delectus."],"created_at":"1994-12-30T14:30:47Z","date":"1996-10-26","date_closed":"1993-08-14","id":12043849106673499505,"responsible_party":"Recusandae quae nobis.","root_cause":"Aut voluptatem deserunt non dolor omnis.","scope":"Qui assumenda vitae reprehenderit dolores.","severity":0,"slack_channel":"Aliquid dolore harum.","summary":"Ipsam id praesentium blanditiis adipisci.","title":"Tempora velit sed minima distinctio velit.","updated_at":"2004-11-14T00:10:42Z"},{"Permissions":"analyst","affected_customers":["Quis ipsam.","Voluptatem eum.","Vero aliquam quisquam veritatis quia quis delectus."],"created_at":"1994-12-30T14:30:47Z","date":"1996-10-26","date_closed":"1993-08-14","id":12043849106673499505,"responsible_party":"Recusandae quae nobis.","root_cause":"Aut voluptatem deserunt non dolor omnis.","scope":"Qui assumenda vitae reprehenderit dolores.","severity":0,"slack_channel":"Aliquid dolore harum.","summary":"Ipsam id praesentium blanditiis adipisci.","title":"Tempora velit sed minima distinctio velit.","updated_at":"2004-11-14T00:10:42Z"}]},"example":[{"Permissions":"analyst","affected_customers":["Quis ipsam.","Voluptatem eum.","Vero aliquam quisquam veritatis quia quis delectus."],"created_at":"1994-12-30T14:30:47Z","date":"1996-10-26","date_closed":"1993-08-14","id":12043849106673499505,"responsible_party":"Recusandae quae nobis.","root_cause":"Aut voluptatem deserunt non dolor omnis.","scope":"Qui assumenda vitae reprehenderit dolores.","severity":0,"slack_channel":"Aliquid dolore harum.","summary":"Ipsam id praesentium blanditiis adipisci.","title":"Tempora velit sed minima distinctio velit.","updated_at":"2004-11-14T00:10:42Z"},{"Permissions":"analyst","affected_customers":["Quis ipsam.","Voluptatem eum.","Vero aliquam quisquam veritatis quia quis delectus."],"created_at":"1994-12-30T14:30:47Z","date":"1996-10-26","date_closed":"1993-08-14","id":12043849106673499505,"responsible_party":"Recusandae quae nobis.","root_cause":"Aut voluptatem deserunt non dolor omnis.","scope":"Qui assumenda vitae reprehenderit dolores.","severity":0,"slack_channel":"Aliquid dolore harum.","summary":"Ipsam id praesentium blanditiis adipisci.","title":"Tempora velit sed minima distinctio velit.","updated_at":"2004-11-14T00:10:42Z"},{"Permissions":"analyst","affected_customers":["Quis ipsam.","Voluptatem eum.","Vero aliquam quisquam veritatis quia quis delectus."],"created_at":"1994-12-30T14:30:47Z","date":"1996-10-26","date_closed":"1993-08-14","id":12043849106673499505,"responsible_party":"Recusandae quae nobis.","root_cause":"Aut voluptatem deserunt non dolor omnis.","scope":"Qui assumenda vitae reprehenderit dolores.","severity":0,"slack_channel":"Aliquid dolore harum.","summary":"Ipsam id praesentium blanditiis adipisci.","title":"Tempora velit sed minima distinctio velit.","updated_at":"2004-11-14T00:10:42Z"}]}}}}}},"/incidents/{id}":{"get":{"tags":["incidents"],"summary":"find incidents","operationId":"incidents#find","parameters":[{"name":"id","in":"path","description":"Unique ID of the incident","required":true,"schema":{"type":"integer","description":"Unique ID of the incident","example":1355286997354828281},"example":1265324971239134687}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Incident"},"example":{"Permissions":"admin","affected_customers":["Aut facere.","Perspiciatis enim.","Dignissimos magnam asperiores ullam dolorem sit ex."],"created_at":"1995-05-02T06:56:39Z","date":"1977-02-11","date_closed":"1982-10-28","id":11941844764433736730,"responsible_party":"Et illo ex aliquid veniam nam facere.","root_cause":"Blanditiis autem.","scope":"Libero voluptatum provident esse odio.","severity":0,"slack_channel":"Et ipsum et eum.","summary":"Ratione porro esse suscipit et.","title":"Totam quia.","updated_at":"1972-11-06T07:23:40Z"}}}}}}},"/vulnerabilities":{"get":{"tags":["vulnerabilities"],"summary":"list vulnerabilities","description":"List all of the vulnerabilities","operationId":"vulnerabilities#list","parameters":[{"name":"limit","in":"query","description":"Limit the number of results","allowEmptyValue":true,"schema":{"type":"integer","description":"Limit the number of results","example":950,"minimum":0,"maximum":1000},"example":374}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Vulnerability"},"example":[{"cvss_score":4.7,"description":"There is a possible out of bounds write due to a use after free. This could lead to remote code execution with no additional execution privileges needed.","exploitable":false,"id":3819473213214139262,"is_patchable":false,"is_upgradeable":false,"title":"CVE-2020-10"},{"cvss_score":4.7,"description":"There is a possible out of bounds write due to a use after free. This could lead to remote code execution with no additional execution privileges needed.","exploitable":false,"id":3819473213214139262,"is_patchable":false,"is_upgradeable":false,"title":"CVE-2020-10"},{"cvss_score":4.7,"description":"There is a possible out of bounds write due to a use after free. This could lead to remote code execution with no additional execution privileges needed.","exploitable":false,"id":3819473213214139262,"is_patchable":false,"is_upgradeable":false,"title":"CVE-2020-10"}]},"example":[{"cvss_score":4.7,"description":"There is a possible out of bounds write due to a use after free. This could lead to remote code execution with no additional execution privileges needed.","exploitable":false,"id":3819473213214139262,"is_patchable":false,"is_upgradeable":false,"title":"CVE-2020-10"},{"cvss_score":4.7,"description":"There is a possible out of bounds write due to a use after free. This could lead to remote code execution with no additional execution privileges needed.","exploitable":false,"id":3819473213214139262,"is_patchable":false,"is_upgradeable":false,"title":"CVE-2020-10"},{"cvss_score":4.7,"description":"There is a possible out of bounds write due to a use after free. This could lead to remote code execution with no additional execution privileges needed.","exploitable":false,"id":3819473213214139262,"is_patchable":false,"is_upgradeable":false,"title":"CVE-2020-10"}]}}}}},"post":{"tags":["vulnerabilities"],"summary":"submit vulnerabilities","operationId":"vulnerabilities#submit","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitRequestBody"},"example":{"cvss_score":0.7719438,"description":"Veniam iste velit illo maxime impedit pariatur.","exploitable":true,"is_patchable":true,"is_upgradeable":true,"title":"Voluptatum ut eveniet."}}}},"responses":{"201":{"description":"Created response."},"404":{"description":"Not Found response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoMatch"},"example":"Vel rerum."}}}}}},"/vulnerabilities/{id}":{"get":{"tags":["vulnerabilities"],"summary":"find vulnerabilities","operationId":"vulnerabilities#find","parameters":[{"name":"id","in":"path","description":"Unique ID of the vulnerability","required":true,"schema":{"type":"integer","description":"Unique ID of the vulnerability","example":8714096609779316103},"example":347206281519856490}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vulnerability"},"example":{"cvss_score":4.7,"description":"There is a possible out of bounds write due to a use after free. This could lead to remote code execution with no additional execution privileges needed.","exploitable":true,"id":11159415604719970017,"is_patchable":true,"is_upgradeable":false,"title":"CVE-2020-10"}}}},"404":{"description":"Not Found response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoMatch"},"example":"Omnis molestiae."}}}}}}},"components":{"schemas":{"Incident":{"type":"object","properties":{"Permissions":{"type":"string","description":"Permissions associated with incident","example":"guest","enum":["admin","analyst","guest"]},"affected_customers":{"type":"array","items":{"type":"string","example":"In autem totam eius atque."},"description":"A list of the affected customers","example":["Non nisi id aut est.","Aut quis aperiam.","Adipisci omnis accusamus qui quibusdam perspiciatis."]},"created_at":{"type":"string","description":"When the incident was submitted","example":"2000-04-06T02:01:12Z","format":"date-time"},"date":{"type":"string","description":"Date the incident occurred","example":"2012-08-28","format":"date"},"date_closed":{"type":"string","description":"Date the incident occurred","example":"2002-07-07","format":"date"},"id":{"type":"integer","description":"Unique ID of the incident","example":525910121978923289,"minimum":0},"responsible_party":{"type":"string","description":"What group or individual caused the initial incident","example":"Enim molestiae vel reprehenderit."},"root_cause":{"type":"string","description":"The original cause of the incident","example":"Ut id eos accusantium."},"scope":{"type":"string","description":"The scope of impact of this incident","example":"Fuga eum sit unde tempora."},"severity":{"type":"integer","description":"The severity of the incident","example":5,"enum":[0,1,2,3,4,5]},"slack_channel":{"type":"string","description":"The slack channel for incident discussions","example":"Mollitia ex."},"summary":{"type":"string","description":"The detailed description of the incident","example":"Fugit iusto nisi."},"title":{"type":"string","description":"The short title of the incident","example":"Quia est et."},"updated_at":{"type":"string","description":"When the incident was last updated","example":"1989-07-27T02:42:12Z","format":"date-time"}},"example":{"Permissions":"analyst","affected_customers":["Quo officia ipsum.","Ut ut aut qui voluptatem.","Et at eaque tempore pariatur."],"created_at":"1970-11-06T00:06:42Z","date":"1982-06-07","date_closed":"1987-08-18","id":13669655108109456858,"responsible_party":"Fugiat repellendus dolor aperiam ad hic qui.","root_cause":"Deleniti odio est corrupti exercitationem occaecati dicta.","scope":"Numquam rerum et molestiae et culpa.","severity":2,"slack_channel":"Nisi veniam.","summary":"Aliquam quo esse ducimus.","title":"Omnis doloribus.","updated_at":"2012-07-01T19:10:09Z"}},"NoMatch":{"type":"string","description":"No vulnerability matched given criteria","example":"Et dolor ipsam."},"SubmitRequestBody":{"type":"object","properties":{"cvss_score":{"type":"number","description":"Severity score of the vulnerability","example":0.5446501,"format":"float"},"description":{"type":"string","description":"Description of the vulnerability","example":"Suscipit provident et quis et."},"exploitable":{"type":"boolean","description":"If the vulnerability is exploitable","example":false},"is_patchable":{"type":"boolean","description":"If the vulnerability is patchable","example":true},"is_upgradeable":{"type":"boolean","description":"If the vulnerability is upgradeable","example":false},"title":{"type":"string","description":"Title of the vulnerability","example":"Adipisci totam hic officia voluptatem nesciunt."}},"example":{"cvss_score":0.8337925,"description":"Similique est dolor sit corporis excepturi quia.","exploitable":false,"is_patchable":true,"is_upgradeable":false,"title":"Placeat modi debitis."}},"Vulnerability":{"type":"object","properties":{"cvss_score":{"type":"number","description":"Severity score of the vulnerability","example":4.7,"format":"float"},"description":{"type":"string","description":"Description of the vulnerability","example":"There is a possible out of bounds write due to a use after free. This could lead to remote code execution with no additional execution privileges needed."},"exploitable":{"type":"boolean","description":"If the vulnerability is exploitable","example":true},"id":{"type":"integer","description":"Unique ID of the vulnerability","example":9125872748072862608,"minimum":0},"is_patchable":{"type":"boolean","description":"If the vulnerability is patchable","example":false},"is_upgradeable":{"type":"boolean","description":"If the vulnerability is upgradeable","example":false},"title":{"type":"string","description":"Title of the vulnerability","example":"CVE-2020-10"}},"example":{"cvss_score":4.7,"description":"There is a possible out of bounds write due to a use after free. This could lead to remote code execution with no additional execution privileges needed.","exploitable":true,"id":1574086337224311383,"is_patchable":true,"is_upgradeable":false,"title":"CVE-2020-10"},"required":["id"]}}}}